Sketch other grouping modes.
