+++ package.json
@@ -262,7 +262,7 @@
                 },
                 {
                     "submenu": "azureResourceGroups.groupBy",
-                    "when": "view == azureResourceGroups",
+                    "when": "view == azureResourceGroups || view == azureResourceGroupsV2",
                     "group": "navigation@2"
                 },
                 {
